moved request body io reader to request's file
now body io reader is contained directly in the request.go file
